+++ src/hooks/MusicDownloadManager.cpp
@@ -10,12 +10,12 @@ class $modify(GDPSMDL, MusicDownloadManager) {
     std::string pathForSFXFolder(int id) {
         if (!GDPSMain::get()->isActive()) return MusicDownloadManager::pathForSFXFolder(id);
         auto server = GDPSMain::get()->m_servers[GDPSMain::get()->m_currentServer];
-        return string::pathToString(dirs::getSaveDir() / "gdpses" / server.saveDir / "");
+        return string::pathToString(dirs::getSaveDir() / "gdpses" / server->saveDir / "");
     }
 
     std::string pathForSongFolder(int id) {
         if (!GDPSMain::get()->isActive()) return MusicDownloadManager::pathForSongFolder(id);
         auto server = GDPSMain::get()->m_servers[GDPSMain::get()->m_currentServer];
-        return string::pathToString(dirs::getSaveDir() / "gdpses" / server.saveDir / "");
+        return string::pathToString(dirs::getSaveDir() / "gdpses" / server->saveDir / "");
     }
 };